% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\title{Estimate Statistics Related to Poisson Distribution}
\usage{
estimatePois(y, lambda, SEtype = c("basic", "bootstrap"), B = 100)
}
\arguments{
\item{y}{The vector of observed data.}

\item{lambda}{The assumed value of \eqn{\lambda}.}

\item{SEtype}{The type of standard errors, which can be either \code{basic} or \code{bootstrap}.}

\item{B}{The number of bootstrapped resamplings, only useful when \code{SEtype = 'bootstrap'}. The default is 100.}
}
\value{
An object of S4 class \code{PoisMLE} with following slots:
\item{y}{The original data.}
\item{MLE}{The maximum likelihood estimator for this dataset.}
\item{LL}{The log likelihood calculated from the observed data assuming the MLE is correct.}
\item{SE}{The standard error for the MLE.}
\item{SEtype}{The method used to calculate the standard error.}
}
\description{
Estimate statistics related to Poisson distribution, including the log likelihood, MLE estimate, and its associated standard error.
}
\examples{
set.seed(123)
y <- sample(x=1:10, size=20, replace=TRUE)
estimatePois(y, lambda = 2, SEtype = "bootstrap")
}
\seealso{
\link[PoisMLE]{logLik}, \link[PoisMLE]{mle}, \link[PoisMLE]{standardError}
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
