% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\title{Calculate the Log Likelihood of Poisson Distribution}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{The vector of observed data.}

\item{lambda}{The assumed value of \eqn{\lambda}.}
}
\value{
A numeric value indicating the log likelihood
}
\description{
Calculate the Log Likelihood of Poisson Distribution
}
\examples{
set.seed(123)
y <- sample(x=1:10, size=20, replace=TRUE)
logLik(y,2)
}
\seealso{
mle, standardError, estimatePois
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
