% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\title{Calculate the MLE Estimator of Poisson Distribution}
\usage{
mle(y)
}
\arguments{
\item{y}{The vector of observed data.}
}
\value{
A numeric value indicating the MLE estimator.
}
\description{
Calculate the MLE estimator of Poisson distribution based on observation data.
}
\examples{
set.seed(123)
y <- sample(x=1:10, size=20, replace=TRUE)
mle(y)
}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{standardError}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
