% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisMLE_class.R
\docType{class}
\name{check_PoisMLE}
\alias{check_PoisMLE}
\alias{poissonMLE}
\alias{PoisMLE-class}
\alias{initialize,PoisMLE-method}
\title{fitted model for PoisMLE}
\usage{
check_PoisMLE(object)

\S4method{initialize}{PoisMLE}(.Object, ...)
}
\description{
Finds a fitted model of class \code{PoisMLE} using the function \code{estimatePois}
}
\details{
The class has the following slots:
\itemize{
\item{y}{The original data}
\item{MLE}{The maximum likelihood estimator for this dataset}
\item{LL}{The log likelihood calculated from the observed data assuming the MLE is correct}
\item{SE}{The standard error for the MLE}
\item{SEtype}{The method used to calculate the standard error}
}
}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{mle}}

\code{\link{standardError}}

\code{\link{plotMLE}}
}
\author{
Annie Jarman
}
