% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE_Class.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\title{Define PoisMLE (S4) Class}
\description{
An Object of class \code{PoisMLE} can be created by the \code{estimatePois()} function
}
\details{
An object of the class `PoisMLE` has the following slots:
\itemize{
\item \code{y} The original data.
\item \code{MLE} The maximum likelihood estimator for this dataset.
\item \code{LL} The log likelihood calculated from the observed data assuming the MLE is correct.
\item \code{SE} The standard error for the MLE.
\item \code{SEtype} The method used to calculate the standard error.
}

The\code{\link[PoisMLE]{estimatePois}} function calculates several statistics 
related to Poisson distribution based on a vector of observed data, including the log likelihood, 
MLE estimate, and its associated standard error. The output is an Object of class \code{PoisMLE}.
One can use the \code{plot()} function to produce a simple visualization of the MLE estimate with 95% confidence interval.
}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{mle}}, \code{\link[PoisMLE]{standardError}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Rex W. Deng: \email{weiye.deng@wustl.edu}
}
