% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePois.R
\name{estimatePois}
\alias{estimatePois}
\alias{estimatePoisson}
\title{estimate Poisson fitted model function}
\usage{
estimatePois(y, SEtype, B = 1000)
}
\arguments{
\item{y}{The vector of observed data}

\item{SEtype}{One of two methods of calculation; basic or bootstrap}

\item{B}{The number of bootstrapped resamplings if using SEtype="bootstrap". Default set to 1000.}
}
\value{
an object of class \code{PoisMLE} containing
\item{y}{The original data of positive integers}
\item{MLE}{The maximum likelihood estimator for this dataset}
\item{LL}{The log likelihood calculated from the observed data assuming the MLE is correct}
\item{SE}{The standard error. For SEtype="bootstrap", this is the standard deviation of bootstrapped MLEs.}
\item{SEtype}{One of two methods - basic or bootstrap - used to calculate the standard error}
}
\description{
Takes in data and returns an object of class \code{PoisMLE}.
}
\seealso{
\code{\link{standardError}}

\code{\link{logLik}}

\code{\link{mle}}
}
\author{
Annie Jarman
}
