% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik}
\alias{logLik}
\alias{logLikelihood}
\title{Log likelihood function}
\usage{
logLik(y, lambda)
}
\arguments{
\item{y}{The vector of observed data}

\item{lambda}{the assumed value of lambda}
}
\value{
Numeric value
\item{LL}{The log likelihood for the observed data conditioned on the value of lamda}
}
\description{
Calculates a log likelihood for the observed data. Used for object of class \code{PoisMLE}.
}
\note{
This is used internally for the \code{estimatePois} function
}
\examples{

set.seed(1227)
y <- rpois(12270,130)
lambda <- mle(y)
logLik(y,lambda)
}
\seealso{
\code{\link{estimatePois}}

\code{\link{mle}}

\code{\link{standardError}}
}
\author{
Annie Jarman
}
