% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle.R
\name{mle}
\alias{mle}
\alias{MLE}
\title{MLE function}
\usage{
mle(y)
}
\arguments{
\item{y}{The vector of observed data}
}
\value{
Numeric value
\item{MLE}{The maximum likelihood estimator for lamda}
}
\description{
Calculates the maximum likelihood estimator for lambda. Used for object of class \code{PoisMLE}.
}
\note{
This is used internally for the \code{estimatePois} function
}
\examples{

set.seed(1227)
y <- rpois(1227,130)
mle(y)
}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{standardError}}
}
\author{
Annie Jarman
}
