% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoisMLE_Class.R
\docType{class}
\name{PoisMLE-class}
\alias{PoisMLE-class}
\title{Define PoisMLE (S4) Class}
\description{
Object of class \code{PoisMLE} created by the \code{estimatePois} functions
}
\details{
An object of the class `PoisMLE` has the following slots:
\itemize{
\item \code{y} The original data.
\item \code{MLE} The maximum likelihood estimator for this dataset.
\item \code{LL} The log likelihood calculated from the observed data assuming the MLE is correct.
\item \code{SE} The standard error for the MLE.
\item \code{SEtype} The method used to calculate the standard error.
}
}
\author{
Rex W. Deng: \email{weiye.deng@wustl.edu}
}
