% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\alias{SE}
\title{Standard Error function}
\usage{
standardError(y, SEtype, B = 1000)
}
\arguments{
\item{y}{The vector of observed data}

\item{SEtype}{One of two methods of calculation; basic or bootstrap}

\item{B}{The number of bootstrapped resamplings if using SEtype="bootstrap". Default set to 1000.}
}
\value{
Numeric value
\item{SE}{The standard error. For SEtype="bootstrap", this is the standard deviation of bootstrapped MLEs.}
}
\description{
Calculates the standard error. Used for object of class \code{PoisMLE}.
}
\note{
This is used internally for the \code{estimatePois} function
}
\examples{

set.seed(1227)
y <- rpois(1227,130)
standardError(y, "basic")
standardError(y, "bootstrap", 1000)
}
\seealso{
\code{\link{estimatePois}}

\code{\link{logLik}}

\code{\link{mle}}
}
\author{
Annie Jarman
}
