% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardError.R
\name{standardError}
\alias{standardError}
\title{Calculate the Standard Error associated with the MLE Estimator (Poisson)}
\usage{
standardError(y, SEtype = c("basic", "bootstrap"), B = 100)
}
\arguments{
\item{y}{The vector of observed data.}

\item{SEtype}{The type of standard errors, which can be either \code{basic} or \code{bootstrap}.}

\item{B}{The number of bootstrapped resamplings, only useful when \code{SEtype = 'bootstrap'}. The default is 100.}
}
\value{
A numeric value indicating the standard error.
}
\description{
Calculate the standard rrror associated with the MLE estimator (Poisson) based on observation data.
}
\examples{
set.seed(123)
y <- sample(x=1:10, size=20, replace=TRUE)
standardError(y, SEtype = 'bootstrap')
}
\seealso{
\code{\link[PoisMLE]{logLik}}, \code{\link[PoisMLE]{mle}}, \code{\link[PoisMLE]{estimatePois}}
}
\author{
Rex W. Deng <\email{weiye.deng@wustl.edu}>
}
